Restore test task as a named task rather than as a named function
After some experimentation, it seems that invoking gulp.series() or
gulp.parallel() within a function is probably an anti-pattern because
Gulp can't build a complete task map without actually executing these
tasks.  For example, the "gulp --tasks" command will not show any
subtasks for those tasks in which gulp.series() or gulp.parallel() are
not invoked when the Gulpfile is loaded.

I chose not to revert to using a variable to store the test task because
it does not show the "test" task when it is a subtask of another task
(e.g. the "default" task).  The reason why gulp.task() seems to work as
desired is because it actually creates a wrapper function around the
task function.  This is what we tried to do with the previous approach,
but the Undertaker library (which manages the Gulp tasks) adds
additional metadata to this wrapper function which makes it work as
desired.

Hopefully before Gulp 4 is released, there will be a way to define a
composite task without using gulp.task() but that behaves just as if it
were defined using gulp.task().