Remove custom socket used to stop server process
This seems like a potential security risk (because it does not require
authentication) and is only used for development.  We avoided a
kill-style solution previously because running babel-node spawns a child
process, and the native Node process.kill() function does not handle
process trees.  However, the tree-kill module does handle killing
process trees, so we revert back to the kill-style solution (which is
cross platform because tree-kill is cross platform).